Fix an issue with filtering UITableView
The table view was being reloaded before the underlying data was
done being filtered, and this was causing the UITableView to have
a mismatch between the number of cells it expected to see (10, as
the filter had not been applied yet) and the number of cells that
actually existed (3, as the filter had been applied once the cells
were being requested).

To fix this, we push the reload of the table view to happen in the
didSet of the filter variable. This ensures that both of the table
view delegate methods will see the same array of data.